converted lock component to screen
