+++ dioxus-wallet/src/components/mod.rs
@@ -2,6 +2,5 @@ pub mod alerts;
 pub mod button;
 pub mod common;
 pub mod loader;
-pub mod lock;
 pub mod token_list;
 pub mod txn_list;
