+++ dioxus-wallet/src/routes/mod.rs
@@ -1,3 +1,4 @@
+mod lock;
 mod recv;
 mod scan;
 mod send;
@@ -5,6 +6,7 @@ mod signup;
 mod tokens;
 mod txns;
 
+pub use lock::Lock;
 pub use recv::Recv;
 pub use scan::Scan;
 pub use send::Send;
